interface working with druid APIs
