+++ pkg/query-service/app/druidReader/reader.go
@@ -1,22 +1,49 @@
 package druidReader
 
 import (
+	"context"
+
 	"go.signoz.io/query-service/druidQuery"
+	"go.signoz.io/query-service/godruid"
 	"go.signoz.io/query-service/model"
 )
 
 type DruidReader struct {
-	Client
-	SqlClient
+	Client    *godruid.Client
+	SqlClient *druidQuery.SqlClient
 }
 
-func NewSpanReader() {
+func NewReader(druidClientUrl string) *DruidReader {
+
 	initialize()
+
+	client := godruid.Client{
+		Url:   druidClientUrl,
+		Debug: true,
+	}
+
+	sqlClient := druidQuery.SqlClient{
+		Url:   druidClientUrl,
+		Debug: true,
+	}
+	return &DruidReader{
+		Client:    &client,
+		SqlClient: &sqlClient,
+	}
+
 }
 func initialize() {
 
 }
 
-func (druid *Druid) GetServices(client, query *model.GetServicesParams) {
-	return druidQuery.GetServices(druid.sqlClient, query)
+func (druid *DruidReader) GetServiceOverview(ctx context.Context, query *model.GetServiceOverviewParams) (*[]model.ServiceOverviewItem, error) {
+	return druidQuery.GetServiceOverview(druid.SqlClient, query)
+}
+
+func (druid *DruidReader) GetServices(ctx context.Context, query *model.GetServicesParams) (*[]model.ServiceItem, error) {
+	return druidQuery.GetServices(druid.SqlClient, query)
+}
+
+func (druid *DruidReader) GetApplicationPercentiles(ctx context.Context, query *model.ApplicationPercentileParams) ([]godruid.Timeseries, error) {
+	return druidQuery.GetApplicationPercentiles(druid.Client, query)
 }
