+++ pkg/query-service/app/http_handler.go
@@ -1,14 +1,13 @@
 package app
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
 
 	"github.com/gorilla/mux"
 	"github.com/posthog/posthog-go"
-	"go.signoz.io/query-service/druidQuery"
-	"go.signoz.io/query-service/godruid"
 	"go.uber.org/zap"
 )
 
@@ -23,17 +22,15 @@ type APIHandler struct {
 	// queryParser  queryParser
 	basePath   string
 	apiPrefix  string
-	client     *godruid.Client
-	sqlClient  *druidQuery.SqlClient
+	reader     *Reader
 	pc         *posthog.Client
 	distinctId string
 }
 
 // NewAPIHandler returns an APIHandler
-func NewAPIHandler(client *godruid.Client, sqlClient *druidQuery.SqlClient, pc *posthog.Client, distinctId string) *APIHandler {
+func NewAPIHandler(reader *Reader, pc *posthog.Client, distinctId string) *APIHandler {
 	aH := &APIHandler{
-		client:     client,
-		sqlClient:  sqlClient,
+		reader:     reader,
 		pc:         pc,
 		distinctId: distinctId,
 	}
@@ -61,20 +58,20 @@ func (aH *APIHandler) RegisterRoutes(router *mux.Router) {
 	router.HandleFunc("/api/v1/user", aH.user).Methods(http.MethodPost)
 	router.HandleFunc("/api/v1/get_percentiles", aH.getApplicationPercentiles).Methods(http.MethodGet)
 	router.HandleFunc("/api/v1/services", aH.getServices).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/services/list", aH.getServicesList).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/services/list", aH.getServicesList).Methods(http.MethodGet)
 	router.HandleFunc("/api/v1/service/overview", aH.getServiceOverview).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/dbOverview", aH.getServiceDBOverview).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/externalAvgDuration", aH.GetServiceExternalAvgDuration).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/externalErrors", aH.getServiceExternalErrors).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/external", aH.getServiceExternal).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/{service}/operations", aH.getOperations).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/service/top_endpoints", aH.getTopEndpoints).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/spans", aH.searchSpans).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/spans/aggregates", aH.searchSpansAggregates).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/tags", aH.searchTags).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/traces/{traceId}", aH.searchTraces).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/usage", aH.getUsage).Methods(http.MethodGet)
-	router.HandleFunc("/api/v1/serviceMapDependencies", aH.serviceMapDependencies).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/dbOverview", aH.getServiceDBOverview).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/externalAvgDuration", aH.GetServiceExternalAvgDuration).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/externalErrors", aH.getServiceExternalErrors).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/external", aH.getServiceExternal).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/{service}/operations", aH.getOperations).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/service/top_endpoints", aH.getTopEndpoints).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/spans", aH.searchSpans).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/spans/aggregates", aH.searchSpansAggregates).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/tags", aH.searchTags).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/traces/{traceId}", aH.searchTraces).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/usage", aH.getUsage).Methods(http.MethodGet)
+	// router.HandleFunc("/api/v1/serviceMapDependencies", aH.serviceMapDependencies).Methods(http.MethodGet)
 }
 
 func (aH *APIHandler) user(w http.ResponseWriter, r *http.Request) {
@@ -102,147 +99,147 @@ func (aH *APIHandler) user(w http.ResponseWriter, r *http.Request) {
 
 }
 
-func (aH *APIHandler) getOperations(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getOperations(w http.ResponseWriter, r *http.Request) {
 
-	vars := mux.Vars(r)
-	serviceName := vars["service"]
+// 	vars := mux.Vars(r)
+// 	serviceName := vars["service"]
 
-	var err error
-	if len(serviceName) == 0 {
-		err = fmt.Errorf("service param not found")
-	}
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	var err error
+// 	if len(serviceName) == 0 {
+// 		err = fmt.Errorf("service param not found")
+// 	}
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetOperations(aH.sqlClient, serviceName)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetOperations(aH.sqlClient, serviceName)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getServicesList(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getServicesList(w http.ResponseWriter, r *http.Request) {
 
-	result, err := druidQuery.GetServicesList(aH.sqlClient)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServicesList(aH.sqlClient)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) searchTags(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) searchTags(w http.ResponseWriter, r *http.Request) {
 
-	serviceName := r.URL.Query().Get("service")
+// 	serviceName := r.URL.Query().Get("service")
 
-	result, err := druidQuery.GetTags(aH.sqlClient, serviceName)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetTags(aH.sqlClient, serviceName)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getTopEndpoints(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getTopEndpoints(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetTopEndpointsRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetTopEndpointsRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetTopEndpoints(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetTopEndpoints(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getUsage(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getUsage(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetUsageRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetUsageRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetUsage(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetUsage(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getServiceDBOverview(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getServiceDBOverview(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetServiceExternalRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetServiceExternalRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetServiceDBOverview(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServiceDBOverview(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getServiceExternal(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getServiceExternal(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetServiceExternalRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetServiceExternalRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetServiceExternal(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServiceExternal(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) GetServiceExternalAvgDuration(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) GetServiceExternalAvgDuration(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetServiceExternalRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetServiceExternalRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetServiceExternalAvgDuration(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServiceExternalAvgDuration(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
-func (aH *APIHandler) getServiceExternalErrors(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) getServiceExternalErrors(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetServiceExternalRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetServiceExternalRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetServiceExternalErrors(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServiceExternalErrors(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
+// }
 
 func (aH *APIHandler) getServiceOverview(w http.ResponseWriter, r *http.Request) {
 
@@ -251,7 +248,7 @@ func (aH *APIHandler) getServiceOverview(w http.ResponseWriter, r *http.Request)
 		return
 	}
 
-	result, err := druidQuery.GetServiceOverview(aH.sqlClient, query)
+	result, err := (*aH.reader).GetServiceOverview(context.Background(), query)
 	if aH.handleError(w, err, http.StatusBadRequest) {
 		return
 	}
@@ -267,7 +264,7 @@ func (aH *APIHandler) getServices(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	result, err := druidQuery.GetServices(aH.sqlClient, query)
+	result, err := (*aH.reader).GetServices(context.Background(), query)
 	if aH.handleError(w, err, http.StatusBadRequest) {
 		return
 	}
@@ -282,63 +279,63 @@ func (aH *APIHandler) getServices(w http.ResponseWriter, r *http.Request) {
 	aH.writeJSON(w, r, result)
 }
 
-func (aH *APIHandler) serviceMapDependencies(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) serviceMapDependencies(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseGetServicesRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseGetServicesRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.GetServiceMapDependencies(aH.sqlClient, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.GetServiceMapDependencies(aH.sqlClient, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
-}
+// 	aH.writeJSON(w, r, result)
+// }
 
-func (aH *APIHandler) searchTraces(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) searchTraces(w http.ResponseWriter, r *http.Request) {
 
-	vars := mux.Vars(r)
-	traceId := vars["traceId"]
+// 	vars := mux.Vars(r)
+// 	traceId := vars["traceId"]
 
-	result, err := druidQuery.SearchTraces(aH.client, traceId)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.SearchTraces(aH.client, traceId)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
+// 	aH.writeJSON(w, r, result)
 
-}
-func (aH *APIHandler) searchSpansAggregates(w http.ResponseWriter, r *http.Request) {
+// }
+// func (aH *APIHandler) searchSpansAggregates(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseSearchSpanAggregatesRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseSearchSpanAggregatesRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.SearchSpansAggregate(aH.client, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.SearchSpansAggregate(aH.client, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
-}
+// 	aH.writeJSON(w, r, result)
+// }
 
-func (aH *APIHandler) searchSpans(w http.ResponseWriter, r *http.Request) {
+// func (aH *APIHandler) searchSpans(w http.ResponseWriter, r *http.Request) {
 
-	query, err := parseSpanSearchRequest(r)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	query, err := parseSpanSearchRequest(r)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	result, err := druidQuery.SearchSpans(aH.client, query)
-	if aH.handleError(w, err, http.StatusBadRequest) {
-		return
-	}
+// 	result, err := druidQuery.SearchSpans(aH.client, query)
+// 	if aH.handleError(w, err, http.StatusBadRequest) {
+// 		return
+// 	}
 
-	aH.writeJSON(w, r, result)
-}
+// 	aH.writeJSON(w, r, result)
+// }
 
 func (aH *APIHandler) getApplicationPercentiles(w http.ResponseWriter, r *http.Request) {
 	// vars := mux.Vars(r)
@@ -348,7 +345,7 @@ func (aH *APIHandler) getApplicationPercentiles(w http.ResponseWriter, r *http.R
 		return
 	}
 
-	result, err := druidQuery.GetApplicationPercentiles(aH.client, query)
+	result, err := (*aH.reader).GetApplicationPercentiles(context.Background(), query)
 	if aH.handleError(w, err, http.StatusBadRequest) {
 		return
 	}
