+++ pkg/query-service/app/server.go
@@ -1,8 +1,10 @@
 package app
 
 import (
+	"fmt"
 	"net"
 	"net/http"
+	"os"
 	"time"
 
 	"github.com/google/uuid"
@@ -11,16 +13,15 @@ import (
 	"github.com/posthog/posthog-go"
 	"github.com/rs/cors"
 	"github.com/soheilhy/cmux"
-	"go.signoz.io/query-service/druidQuery"
-	"go.signoz.io/query-service/godruid"
+	"go.signoz.io/query-service/app/druidReader"
 	"go.signoz.io/query-service/healthcheck"
 	"go.signoz.io/query-service/utils"
 	"go.uber.org/zap"
 )
 
 type ServerOptions struct {
-	HTTPHostPort   string
-	DruidClientUrl string
+	HTTPHostPort string
+	// DruidClientUrl string
 }
 
 // Server runs HTTP, Mux and a grpc server
@@ -28,11 +29,10 @@ type Server struct {
 	// logger       *zap.Logger
 	// querySvc     *querysvc.QueryService
 	// queryOptions *QueryOptions
-	serverOptions *ServerOptions
 
 	// tracer opentracing.Tracer // TODO make part of flags.Service
-
-	conn net.Listener
+	serverOptions *ServerOptions
+	conn          net.Listener
 	// grpcConn           net.Listener
 	httpConn net.Listener
 	// grpcServer         *grpc.Server
@@ -64,6 +64,11 @@ func NewServer(serverOptions *ServerOptions) (*Server, error) {
 	// if err != nil {
 	// 	return nil, err
 	// }
+	httpServer, err := createHTTPServer()
+
+	if err != nil {
+		return nil, err
+	}
 
 	return &Server{
 		// logger: logger,
@@ -72,7 +77,7 @@ func NewServer(serverOptions *ServerOptions) (*Server, error) {
 		// tracer:             tracer,
 		// grpcServer:         grpcServer,
 		serverOptions: serverOptions,
-		httpServer:    createHTTPServer(serverOptions.DruidClientUrl),
+		httpServer:    httpServer,
 		separatePorts: true,
 		// separatePorts:      grpcPort != httpPort,
 		unavailableChannel: make(chan healthcheck.Status),
@@ -82,22 +87,25 @@ func NewServer(serverOptions *ServerOptions) (*Server, error) {
 var posthogClient posthog.Client
 var distinctId string
 
-func createHTTPServer(druidClientUrl string) *http.Server {
+func createHTTPServer() (*http.Server, error) {
 
 	posthogClient = posthog.New("H-htDCae7CR3RV57gUzmol6IAKtm5IMCvbcm_fwnL-w")
 	distinctId = uuid.New().String()
 
-	client := godruid.Client{
-		Url:   druidClientUrl,
-		Debug: true,
+	var reader Reader
+
+	storage := os.Getenv("STORAGE")
+	if storage == "druid" {
+		druidClientUrl := os.Getenv("DruidClientUrl")
+		reader = druidReader.NewReader(druidClientUrl)
+	} else if storage == "clickhouse" {
+		// clickHouseClientUrl := os.Getenv("clickHouseClientUrl")
+		// reader = clickHouseReader.NewTraceReader()
+	} else {
+		return nil, fmt.Errorf("Storage type: %s is not supported in query service", storage)
 	}
 
-	sqlClient := druidQuery.SqlClient{
-		Url:   druidClientUrl,
-		Debug: true,
-	}
-
-	apiHandler := NewAPIHandler(&client, &sqlClient, &posthogClient, distinctId)
+	apiHandler := NewAPIHandler(&reader, &posthogClient, distinctId)
 	r := NewRouter()
 
 	r.Use(analyticsMiddleware)
@@ -118,7 +126,7 @@ func createHTTPServer(druidClientUrl string) *http.Server {
 
 	return &http.Server{
 		Handler: handler,
-	}
+	}, nil
 }
 
 func loggingMiddleware(next http.Handler) http.Handler {
